Update signable_starknet.go
I have modified one comment in OrderPayload class to say that price is scaled by 1e8.  Comments are now aligned with .go code sample we have in /benchmarks.

This was highlighted by one of our customers, that information about how price should be formatted was unclear in the original code example.